feat: add edit agent
