+++ app/javascript/dashboard/routes/dashboard/settings/agents/Index.vue
@@ -75,16 +75,6 @@
                         :is-loading="loading[agent.id]"
                         @click="openDeletePopup(agent, index)"
                       />
-                      <woot-button
-                        v-if="showEditPermissionsAction(agent)"
-                        v-tooltip.top="$t('AGENT_MGMT.PERMISSIONS.BUTTON_TEXT')"
-                        variant="smooth"
-                        color-scheme="secondary"
-                        size="tiny"
-                        icon="lock-closed"
-                        class-names="grey-btn"
-                        @click="openEditPermissions(agent, index)"
-                      />
                     </div>
                   </td>
                 </tr>
@@ -131,17 +121,6 @@
       :confirm-text="deleteConfirmText"
       :reject-text="deleteRejectText"
     />
-    <!-- Edit Permissions -->
-    <woot-modal
-      :show.sync="showEditPermissions"
-      :on-close="closeEditPermissions"
-    >
-      <edit-permission-agent
-        v-if="showEditPermissions"
-        :agent="currentAgent"
-        :on-close="closeEditPermissions"
-      />
-    </woot-modal>
   </div>
 </template>
 <script>
@@ -150,14 +129,12 @@ import globalConfigMixin from 'shared/mixins/globalConfigMixin';
 import Thumbnail from '../../../../components/widgets/Thumbnail.vue';
 import AddAgent from './AddAgent.vue';
 import EditAgent from './EditAgent.vue';
-import EditPermissionAgent from './EditPermissionAgent.vue';
 
 export default {
   components: {
     AddAgent,
     EditAgent,
     Thumbnail,
-    EditPermissionAgent,
   },
   mixins: [globalConfigMixin],
   data() {
@@ -166,7 +143,6 @@ export default {
       showAddPopup: false,
       showDeletePopup: false,
       showEditPopup: false,
-      showEditPermissions: false,
       agentAPI: {
         message: '',
       },
@@ -215,12 +191,7 @@ export default {
       }
       return true;
     },
-    showEditPermissionsAction(agent) {
-      if (agent.role === 'administrator') {
-        return this.verifiedAdministrators().length !== 1;
-      }
-      return true;
-    },
+
     verifiedAdministrators() {
       return this.agentList.filter(
         agent => agent.role === 'administrator' && agent.confirmed
@@ -250,15 +221,6 @@ export default {
     },
     closeDeletePopup() {
       this.showDeletePopup = false;
-    },
-
-    // Edit Permission Function
-    openEditPermissions(agent) {
-      this.showEditPermissions = true;
-      this.currentAgent = agent;
-    },
-    closeEditPermissions() {
-      this.showEditPermissions = false;
     },
 
     confirmDeletion() {
