+++ db/migrate/20240614014644_add_permissions_to_account_users.rb
@@ -1,5 +1,12 @@
 class AddPermissionsToAccountUsers < ActiveRecord::Migration[7.0]
   def change
-    add_column :account_users, :permissions, :jsonb, default: {}
+    add_column :account_users, :permissions, :jsonb, default: {
+      'conversations': true,
+      'accounts': true,
+      'teams': true,
+      'reports': true,
+      'contacts': true,
+      'labels': true
+    }
   end
 end
